Move out dependency calculation to a dedicated module
