+++ crates/gitbutler-branch-actions/src/lib.rs
@@ -31,6 +31,8 @@ pub use branch_manager::{BranchManager, BranchManagerExt};
 mod base;
 pub use base::BaseBranch;
 
+mod dependencies;
+
 pub mod upstream_integration;
 
 mod integration;
